# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  lessondata.VideoResources:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      url:
        type: string
    type: object
...
  questionBank.Programm:
    properties:
      canExam:
        type: integer
      canPractice:
        type: integer
      chapterId:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      describe:
        type: string
      id:
        description: 主键ID
        type: integer
      problemType:
        type: integer
      title:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
  questionBank.ProgrammCase:
    properties:
      clockLimit:
        type: integer
      cpuLimit:
        type: integer
      cpuRateLimit:
        type: integer
      cpuSetLimit:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      input:
        type: string
      languageId:
        type: integer
      memoryLimit:
        type: integer
      name:
        type: string
      output:
        type: string
      procLimit:
        type: integer
      programmId:
        type: integer
      score:
        type: integer
      stackLimit:
        type: integer
      strictMemoryLimit:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  request.MakeProgramm:
    properties:
      canExam:
        type: integer
      canPractice:
        type: integer
      chapterId:
        type: integer
      createdAt:
        description: 创建时间
        type: string
      describe:
        type: string
      id:
        description: 主键ID
        type: integer
      problemType:
        type: integer
      support_language:
        items:
          $ref: '#/definitions/request.SupportLanguage'
        type: array
      title:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.SupportLanguage:
    properties:
      cases:
        items:
          $ref: '#/definitions/questionBank.ProgrammCase'
        type: array
      createdAt:
        description: 创建时间
        type: string
      defaultCode:
        type: string
      id:
        description: 主键ID
        type: integer
      languageId:
        type: integer
      programmId:
        type: integer
      referenceAnswer:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /questionBankProgramm/createQuestionBankProgramm:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建QuestionBankProgramm
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MakeProgramm'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建QuestionBankProgramm
      tags:
      - Programm
  /questionBankProgramm/deleteQuestionBankProgramm:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除QuestionBankProgramm
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/questionBank.Programm'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除QuestionBankProgramm
      tags:
      - Programm
  /questionBankProgramm/deleteQuestionBankProgrammByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除QuestionBankProgramm
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除QuestionBankProgramm
      tags:
      - Programm
  /questionBankProgramm/findQuestionBankProgramm:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: canExam
        type: integer
      - in: query
        name: canPractice
        type: integer
      - in: query
        name: chapterId
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: describe
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - in: query
        name: problemType
        type: integer
      - in: query
        name: title
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询QuestionBankProgramm
      tags:
      - Programm
  /questionBankProgramm/getQuestionBankProgrammList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: canExam
        type: integer
      - in: query
        name: canPractice
        type: integer
      - in: query
        name: chapterId
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: describe
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: problemType
        type: integer
      - in: query
        name: title
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取QuestionBankProgramm列表
      tags:
      - Programm
  /questionBankProgramm/updateQuestionBankProgramm:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新QuestionBankProgramm
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/questionBank.Programm'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新QuestionBankProgramm
      tags:
      - Programm
...
  /videoResources/createVideoResources:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建VideoResources
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/lessondata.VideoResources'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建VideoResources
      tags:
      - VideoResources
  /videoResources/deleteVideoResources:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除VideoResources
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/lessondata.VideoResources'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除VideoResources
      tags:
      - VideoResources
  /videoResources/deleteVideoResourcesByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除VideoResources
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除VideoResources
      tags:
      - VideoResources
  /videoResources/findVideoResources:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: url
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
  /videoResources/getVideoResourcesList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: url
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取VideoResources列表
      tags:
      - VideoResources
  /videoResources/updateVideoResources:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新VideoResources
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/lessondata.VideoResources'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新VideoResources
      tags:
      - VideoResources
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"